TriCorp Tech_Health_System/requirements.txt
fastapi
uvicorn
pandas
snowflake-connector-python
python-dotenv
